refactor(client): 패키지 이름을 Kakao API를 사용함을 명시한다.
