--- clients/location-client/src/main/java/com/giwankim/taxiservice/client/location/LocationRequest.java
@@ -1,4 +1,0 @@
-package com.giwankim.taxiservice.client.location;
-
-public record LocationRequest(String query, Integer page, Integer size) {
-}
