+++ clients/location-client/src/main/java/com/giwankim/taxiservice/client/location/kakao/KakaoLocationApi.java
@@ -1,4 +1,4 @@
-package com.giwankim.taxiservice.client.location;
+package com.giwankim.taxiservice.client.location.kakao;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.cloud.openfeign.SpringQueryMap;
@@ -6,11 +6,10 @@ import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-@FeignClient(value = "location-kakao-api", url = "${location.kakao.api.url}")
-public interface LocationApi {
+@FeignClient(value = "kakao-location-api", url = "${location.kakao.api.url}")
+interface KakaoLocationApi {
 
   @RequestMapping(method = RequestMethod.GET, value = "/search/address.json",
     consumes = MediaType.APPLICATION_JSON_VALUE)
-  LocationsResponse searchLocations(@SpringQueryMap LocationRequest request);
-
+  KakaoLocationResponse searchLocations(@SpringQueryMap KakaoLocationRequest request);
 }
