+++ clients/location-client/src/main/java/com/giwankim/taxiservice/client/location/kakao/KakaoLocationRequest.java
@@ -1,0 +1,4 @@
+package com.giwankim.taxiservice.client.location.kakao;
+
+record KakaoLocationRequest(String query, Integer page, Integer size) {
+}
